Changes to try to separate out bam functions
