+++ source/SequenceContainer.py
@@ -1,8 +1,6 @@
 import random
 import copy
-import pathlib
 import bisect
-import pickle
 import sys
 
 import numpy as np
@@ -10,84 +8,11 @@ from Bio.Seq import Seq
 
 from source.neat_cigar import CigarString
 from source.probability import DiscreteDistribution, poisson_list
+from source.constants_and_models import ALLOWED_NUCL, DEFAULT_MODEL_1, ALL_IND, COV_FRAGLEN_PERCENTILE, TRI_IND
+from source.constants_and_models import LARGE_NUMBER, MAX_MUTFRAC, IGNORE_TRINUC, MAX_ATTEMPTS, NUC_IND
 
 # TODO This whole file is in desperate need of refactoring
 # TODO 1st step will be to make all cigar strings optional, since that is only needed for bams
-
-"""
-Constants needed for analysis
-"""
-MAX_ATTEMPTS = 100  # max attempts to insert a mutation into a valid position
-MAX_MUTFRAC = 0.3  # the maximum percentage of a window that can contain mutations
-
-NUCL = ['A', 'C', 'G', 'T']
-TRI_IND = {'AA': 0, 'AC': 1, 'AG': 2, 'AT': 3, 'CA': 4, 'CC': 5, 'CG': 6, 'CT': 7,
-           'GA': 8, 'GC': 9, 'GG': 10, 'GT': 11, 'TA': 12, 'TC': 13, 'TG': 14, 'TT': 15}
-NUC_IND = {'A': 0, 'C': 1, 'G': 2, 'T': 3}
-ALL_TRI = [NUCL[i] + NUCL[j] + NUCL[k] for i in range(len(NUCL)) for j in range(len(NUCL)) for k in range(len(NUCL))]
-ALL_IND = {ALL_TRI[i]: i for i in range(len(ALL_TRI))}
-
-# DEBUG
-IGNORE_TRINUC = False
-
-# percentile resolution used for fraglen quantizing
-COV_FRAGLEN_PERCENTILE = 10.
-LARGE_NUMBER = 9999999999
-
-"""
-DEFAULT MUTATION MODELS
-"""
-
-DEFAULT_1_OVERALL_MUT_RATE = 0.001
-DEFAULT_1_HOMOZYGOUS_FREQ = 0.010
-DEFAULT_1_INDEL_FRACTION = 0.05
-DEFAULT_1_INS_VS_DEL = 0.6
-DEFAULT_1_INS_LENGTH_VALUES = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
-DEFAULT_1_INS_LENGTH_WEIGHTS = [0.4, 0.2, 0.1, 0.05, 0.05, 0.05, 0.05, 0.034, 0.033, 0.033]
-DEFAULT_1_DEL_LENGTH_VALUES = [1, 2, 3, 4, 5]
-DEFAULT_1_DEL_LENGTH_WEIGHTS = [0.3, 0.2, 0.2, 0.2, 0.1]
-example_matrix_1 = [[0.0, 0.15, 0.7, 0.15],
-                    [0.15, 0.0, 0.15, 0.7],
-                    [0.7, 0.15, 0.0, 0.15],
-                    [0.15, 0.7, 0.15, 0.0]]
-DEFAULT_1_TRI_FREQS = [copy.deepcopy(example_matrix_1) for _ in range(16)]
-DEFAULT_1_TRINUC_BIAS = [1. / float(len(ALL_TRI)) for _ in ALL_TRI]
-DEFAULT_MODEL_1 = [DEFAULT_1_OVERALL_MUT_RATE,
-                   DEFAULT_1_HOMOZYGOUS_FREQ,
-                   DEFAULT_1_INDEL_FRACTION,
-                   DEFAULT_1_INS_VS_DEL,
-                   DEFAULT_1_INS_LENGTH_VALUES,
-                   DEFAULT_1_INS_LENGTH_WEIGHTS,
-                   DEFAULT_1_DEL_LENGTH_VALUES,
-                   DEFAULT_1_DEL_LENGTH_WEIGHTS,
-                   DEFAULT_1_TRI_FREQS,
-                   DEFAULT_1_TRINUC_BIAS]
-
-DEFAULT_2_OVERALL_MUT_RATE = 0.002
-DEFAULT_2_HOMOZYGOUS_FREQ = 0.200
-DEFAULT_2_INDEL_FRACTION = 0.1
-DEFAULT_2_INS_VS_DEL = 0.3
-DEFAULT_2_INS_LENGTH_VALUES = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
-DEFAULT_2_INS_LENGTH_WEIGHTS = [0.1, 0.1, 0.2, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05]
-# noinspection DuplicatedCode
-DEFAULT_2_DEL_LENGTH_VALUES = [1, 2, 3, 4, 5]
-DEFAULT_2_DEL_LENGTH_WEIGHTS = [0.3, 0.2, 0.2, 0.2, 0.1]
-example_matrix_2 = [[0.0, 0.15, 0.7, 0.15],
-                    [0.15, 0.0, 0.15, 0.7],
-                    [0.7, 0.15, 0.0, 0.15],
-                    [0.15, 0.7, 0.15, 0.0]]
-DEFAULT_2_TRI_FREQS = [copy.deepcopy(example_matrix_2) for _ in range(16)]
-DEFAULT_2_TRINUC_BIAS = [1. / float(len(ALL_TRI)) for _ in ALL_TRI]
-DEFAULT_MODEL_2 = [DEFAULT_2_OVERALL_MUT_RATE,
-                   DEFAULT_2_HOMOZYGOUS_FREQ,
-                   DEFAULT_2_INDEL_FRACTION,
-                   DEFAULT_2_INS_VS_DEL,
-                   DEFAULT_2_INS_LENGTH_VALUES,
-                   DEFAULT_2_INS_LENGTH_WEIGHTS,
-                   DEFAULT_2_DEL_LENGTH_VALUES,
-                   DEFAULT_2_DEL_LENGTH_WEIGHTS,
-                   DEFAULT_2_TRI_FREQS,
-                   DEFAULT_2_TRINUC_BIAS]
 
 
 class SequenceContainer:
@@ -164,8 +89,8 @@ class SequenceContainer:
                                 DiscreteDistribution(n[7], n[6]), []])
             for m in n[8]:
                 # noinspection PyTypeChecker
-                self.models[-1][6].append([DiscreteDistribution(m[0], NUCL), DiscreteDistribution(m[1], NUCL),
-                                           DiscreteDistribution(m[2], NUCL), DiscreteDistribution(m[3], NUCL)])
+                self.models[-1][6].append([DiscreteDistribution(m[0], ALLOWED_NUCL), DiscreteDistribution(m[1], ALLOWED_NUCL),
+                                           DiscreteDistribution(m[2], ALLOWED_NUCL), DiscreteDistribution(m[3], ALLOWED_NUCL)])
             self.models[-1].append([m for m in n[9]])
 
         # initialize poisson attributes
@@ -240,8 +165,8 @@ class SequenceContainer:
                                 DiscreteDistribution(n[7], n[6]), []])
             for m in n[8]:
                 # noinspection PyTypeChecker
-                self.models[-1][6].append([DiscreteDistribution(m[0], NUCL), DiscreteDistribution(m[1], NUCL),
-                                           DiscreteDistribution(m[2], NUCL), DiscreteDistribution(m[3], NUCL)])
+                self.models[-1][6].append([DiscreteDistribution(m[0], ALLOWED_NUCL), DiscreteDistribution(m[1], ALLOWED_NUCL),
+                                           DiscreteDistribution(m[2], ALLOWED_NUCL), DiscreteDistribution(m[3], ALLOWED_NUCL)])
             self.models[-1].append([m for m in n[9]])
 
     def update_trinuc_bias(self):
@@ -510,7 +435,7 @@ class SequenceContainer:
                 if random.random() <= self.models[i][3]:
                     in_len = self.models[i][4].sample()
                     # sequence content of random insertions is uniformly random (change this later, maybe)
-                    in_seq = ''.join([random.choice(NUCL) for _ in range(in_len)])
+                    in_seq = ''.join([random.choice(ALLOWED_NUCL) for _ in range(in_len)])
                     ref_nucl = self.sequences[i][event_pos]
                     my_indel = (event_pos, ref_nucl, ref_nucl + in_seq)
                 # deletion
@@ -847,278 +772,3 @@ class SequenceContainer:
         # read_out[i] = (pos, cigar, read_string, qual_string)
         return read_out
 
-
-class ReadContainer:
-    """
-    Container for read data: computes quality scores and positions to insert errors
-    """
-
-    def __init__(self, read_len, error_model, rescaled_error, rescale_qual=False):
-
-        self.read_len = read_len
-        self.rescale_qual = rescale_qual
-
-        model_path = pathlib.Path(error_model)
-        try:
-            error_dat = pickle.load(open(model_path, 'rb'), encoding="bytes")
-        except IOError:
-            print("\nProblem opening the sequencing error model.\n")
-            sys.exit(1)
-
-        self.uniform = False
-
-        # uniform-error SE reads (e.g., PacBio)
-        if len(error_dat) == 4:
-            self.uniform = True
-            [q_scores, off_q, avg_error, error_params] = error_dat
-            self.uniform_q_score = min([max(q_scores), int(-10. * np.log10(avg_error) + 0.5)])
-            print('Reading in uniform sequencing error model... (q=' + str(self.uniform_q_score) + '+' + str(
-                off_q) + ', p(err)={0:0.2f}%)'.format(100. * avg_error))
-
-        # only 1 q-score model present, use same model for both strands
-        elif len(error_dat) == 6:
-            [init_q1, prob_q1, q_scores, off_q, avg_error, error_params] = error_dat
-            self.pe_models = False
-
-        # found a q-score model for both forward and reverse strands
-        elif len(error_dat) == 8:
-            [init_q1, prob_q1, init_q2, prob_q2, q_scores, off_q, avg_error, error_params] = error_dat
-            self.pe_models = True
-            if len(init_q1) != len(init_q2) or len(prob_q1) != len(prob_q2):
-                print('\nError: R1 and R2 quality score models are of different length.\n')
-                sys.exit(1)
-
-        # This serves as a sanity check for the input model
-        else:
-            print('\nError: Something wrong with error model.\n')
-            sys.exit(1)
-
-        self.q_err_rate = [0.] * (max(q_scores) + 1)
-        for q in q_scores:
-            self.q_err_rate[q] = 10. ** (-q / 10.)
-        self.off_q = off_q
-        self.err_p = error_params
-        # Selects a new nucleotide based on the error model
-        self.err_sse = [DiscreteDistribution(n, NUCL) for n in self.err_p[0]]
-        # allows for selection of indel length based on the parameters of the model
-        self.err_sie = DiscreteDistribution(self.err_p[2], self.err_p[3])
-        # allows for indel insertion based on the length above and the probability from the model
-        self.err_sin = DiscreteDistribution(self.err_p[5], NUCL)
-
-        # adjust sequencing error frequency to match desired rate
-        if rescaled_error is None:
-            self.error_scale = 1.0
-        else:
-            self.error_scale = rescaled_error / avg_error
-            if not self.rescale_qual:
-                print('Warning: Quality scores no longer exactly representative of error probability. '
-                      'Error model scaled by {0:.3f} to match desired rate...'.format(self.error_scale))
-            if self.uniform:
-                if rescaled_error <= 0.:
-                    self.uniform_q_score = max(q_scores)
-                else:
-                    self.uniform_q_score = min([max(q_scores), int(-10. * np.log10(rescaled_error) + 0.5)])
-                print(' - Uniform quality score scaled to match specified error rate (q=' + str(
-                    self.uniform_qscore) + '+' + str(self.off_q) + ', p(err)={0:0.2f}%)'.format(100. * rescaled_error))
-
-        if not self.uniform:
-            # adjust length to match desired read length
-            if self.read_len == len(init_q1):
-                self.q_ind_remap = range(self.read_len)
-            else:
-                print('Warning: Read length of error model (' + str(len(init_q1)) + ') does not match -R value (' + str(
-                    self.read_len) + '), rescaling model...')
-                self.q_ind_remap = [max([1, len(init_q1) * n // read_len]) for n in range(read_len)]
-
-            # initialize probability distributions
-            self.init_dist_by_pos_1 = [DiscreteDistribution(init_q1[i], q_scores) for i in range(len(init_q1))]
-            self.prob_dist_by_pos_by_prev_q1 = [None]
-            for i in range(1, len(init_q1)):
-                self.prob_dist_by_pos_by_prev_q1.append([])
-                for j in range(len(init_q1[0])):
-                    # if we don't have sufficient data for a transition, use the previous quality score
-                    if np.sum(prob_q1[i][j]) <= 0.:
-                        self.prob_dist_by_pos_by_prev_q1[-1].append(
-                            DiscreteDistribution([1], [q_scores[j]], degenerate_val=q_scores[j]))
-                    else:
-                        self.prob_dist_by_pos_by_prev_q1[-1].append(DiscreteDistribution(prob_q1[i][j], q_scores))
-
-            # If paired-end, initialize probability distributions for the other strand
-            if self.pe_models:
-                self.init_dist_by_pos_2 = [DiscreteDistribution(init_q2[i], q_scores) for i in range(len(init_q2))]
-                self.prob_dist_by_pos_by_prev_q2 = [None]
-                for i in range(1, len(init_q2)):
-                    self.prob_dist_by_pos_by_prev_q2.append([])
-                    for j in range(len(init_q2[0])):
-                        if np.sum(prob_q2[i][
-                                      j]) <= 0.:  # if we don't have sufficient data for a transition, use the previous qscore
-                            self.prob_dist_by_pos_by_prev_q2[-1].append(
-                                DiscreteDistribution([1], [q_scores[j]], degenerate_val=q_scores[j]))
-                        else:
-                            self.prob_dist_by_pos_by_prev_q2[-1].append(DiscreteDistribution(prob_q2[i][j], q_scores))
-
-    def get_sequencing_errors(self, read_data, is_reverse_strand=False):
-        """
-        Inserts errors of type substitution, insertion, or deletion into read_data, and assigns a quality score
-        based on the container model.
-
-        :param read_data: sequence to insert errors into
-        :param is_reverse_strand: whether to treat this as the reverse strand or not
-        :return: modified sequence and associate quality scores
-        """
-
-        # TODO this is one of the slowest methods in the code. Need to investigate how to speed this up.
-        q_out = [0] * self.read_len
-        s_err = []
-
-        if self.uniform:
-            my_q = [self.uniform_q_score + self.off_q] * self.read_len
-            q_out = ''.join([chr(n) for n in my_q])
-            for i in range(self.read_len):
-                if random.random() < self.error_scale * self.q_err_rate[self.uniform_q_score]:
-                    s_err.append(i)
-        else:
-            if self.pe_models and is_reverse_strand:
-                my_q = self.init_dist_by_pos_2[0].sample()
-            else:
-                my_q = self.init_dist_by_pos_1[0].sample()
-            q_out[0] = my_q
-
-            # Every time this is hit, we loop the entire read length twice. I feel like these two loops
-            # Could be combined into one fairly easily. The culprit seems to bee too many hits to the sample() method.
-            for i in range(1, self.read_len):
-                if self.pe_models and is_reverse_strand:
-                    my_q = self.prob_dist_by_pos_by_prev_q2[self.q_ind_remap[i]][my_q].sample()
-                else:
-                    my_q = self.prob_dist_by_pos_by_prev_q1[self.q_ind_remap[i]][my_q].sample()
-                q_out[i] = my_q
-
-            if is_reverse_strand:
-                q_out = q_out[::-1]
-
-            for i in range(self.read_len):
-                if random.random() < self.error_scale * self.q_err_rate[q_out[i]]:
-                    s_err.append(i)
-
-            if self.rescale_qual:  # do we want to rescale qual scores to match rescaled error?
-                q_out = [max([0, int(-10. * np.log10(self.error_scale * self.q_err_rate[n]) + 0.5)]) for n in q_out]
-                q_out = [min([int(self.q_err_rate[-1]), n]) for n in q_out]
-                q_out = ''.join([chr(n + self.off_q) for n in q_out])
-            else:
-                q_out = ''.join([chr(n + self.off_q) for n in q_out])
-
-        if self.error_scale == 0.0:
-            return q_out, []
-
-        s_out = []
-        n_del_so_far = 0
-        # don't allow indel errors to occur on subsequent positions
-        prev_indel = -2
-        # don't allow other sequencing errors to occur on bases removed by deletion errors
-        del_blacklist = []
-
-        # Need to check into this loop, to make sure it isn't slowing us down.
-        # The culprit seems to bee too many hits to the sample() method. This has a few of those calls.
-        for ind in s_err[::-1]:  # for each error that we're going to insert...
-
-            # determine error type
-            is_sub = True
-            if ind != 0 and ind != self.read_len - 1 - max(self.err_p[3]) and abs(ind - prev_indel) > 1:
-                if random.random() < self.err_p[1]:
-                    is_sub = False
-
-            # insert substitution error
-            if is_sub:
-                my_nucl = str(read_data[ind])
-                new_nucl = self.err_sse[NUC_IND[my_nucl]].sample()
-                s_out.append(('S', 1, ind, my_nucl, new_nucl))
-
-            # insert indel error
-            else:
-                indel_len = self.err_sie.sample()
-
-                # insertion error
-                if random.random() < self.err_p[4]:
-                    my_nucl = str(read_data[ind])
-                    new_nucl = my_nucl + ''.join([self.err_sin.sample() for n in range(indel_len)])
-                    s_out.append(('I', len(new_nucl) - 1, ind, my_nucl, new_nucl))
-
-                # deletion error (prevent too many of them from stacking up)
-                elif ind < self.read_len - 2 - n_del_so_far:
-                    my_nucl = str(read_data[ind:ind + indel_len + 1])
-                    new_nucl = str(read_data[ind])
-                    n_del_so_far += len(my_nucl) - 1
-                    s_out.append(('D', len(my_nucl) - 1, ind, my_nucl, new_nucl))
-                    for i in range(ind + 1, ind + indel_len + 1):
-                        del_blacklist.append(i)
-                prev_indel = ind
-
-        # remove blacklisted errors
-        for i in range(len(s_out) - 1, -1, -1):
-            if s_out[i][2] in del_blacklist:
-                del s_out[i]
-
-        return q_out, s_out
-
-
-# parse mutation model pickle file
-def parse_input_mutation_model(model=None, which_default=1):
-    if which_default == 1:
-        out_model = [copy.deepcopy(n) for n in DEFAULT_MODEL_1]
-    elif which_default == 2:
-        out_model = [copy.deepcopy(n) for n in DEFAULT_MODEL_2]
-    else:
-        print('\nError: Unknown default mutation model specified\n')
-        sys.exit(1)
-
-    if model is not None:
-        pickle_dict = pickle.load(open(model, "rb"))
-        out_model[0] = pickle_dict['AVG_MUT_RATE']
-        out_model[2] = 1. - pickle_dict['SNP_FREQ']
-
-        ins_list = pickle_dict['INDEL_FREQ']
-        if len(ins_list):
-            ins_count = sum([ins_list[k] for k in ins_list.keys() if k >= 1])
-            del_count = sum([ins_list[k] for k in ins_list.keys() if k <= -1])
-            ins_vals = [k for k in sorted(ins_list.keys()) if k >= 1]
-            ins_weight = [ins_list[k] / float(ins_count) for k in ins_vals]
-            del_vals = [k for k in sorted([abs(k) for k in ins_list.keys() if k <= -1])]
-            del_weight = [ins_list[-k] / float(del_count) for k in del_vals]
-        else:  # degenerate case where no indel stats are provided
-            ins_count = 1
-            del_count = 1
-            ins_vals = [1]
-            ins_weight = [1.0]
-            del_vals = [1]
-            del_weight = [1.0]
-        out_model[3] = ins_count / float(ins_count + del_count)
-        out_model[4] = ins_vals
-        out_model[5] = ins_weight
-        out_model[6] = del_vals
-        out_model[7] = del_weight
-
-        trinuc_trans_prob = pickle_dict['TRINUC_TRANS_PROBS']
-        for k in sorted(trinuc_trans_prob.keys()):
-            my_ind = TRI_IND[k[0][0] + k[0][2]]
-            (k1, k2) = (NUC_IND[k[0][1]], NUC_IND[k[1][1]])
-            out_model[8][my_ind][k1][k2] = trinuc_trans_prob[k]
-        for i in range(len(out_model[8])):
-            for j in range(len(out_model[8][i])):
-                for l in range(len(out_model[8][i][j])):
-                    # if trinuc not present in input mutation model, assign it uniform probability
-                    if float(sum(out_model[8][i][j])) < 1e-12:
-                        out_model[8][i][j] = [0.25, 0.25, 0.25, 0.25]
-                    else:
-                        out_model[8][i][j][l] /= float(sum(out_model[8][i][j]))
-
-        trinuc_mut_prob = pickle_dict['TRINUC_MUT_PROB']
-        which_have_we_seen = {n: False for n in ALL_TRI}
-        trinuc_mean = np.mean(list(trinuc_mut_prob.values()))
-        for trinuc in trinuc_mut_prob.keys():
-            out_model[9][ALL_IND[trinuc]] = trinuc_mut_prob[trinuc]
-            which_have_we_seen[trinuc] = True
-        for trinuc in which_have_we_seen.keys():
-            if not which_have_we_seen[trinuc]:
-                out_model[9][ALL_IND[trinuc]] = trinuc_mean
-
-    return out_model
