+++ source/ref_func.py
@@ -5,10 +5,8 @@ import gzip
 import pathlib
 import random
 from Bio.Seq import Seq
-from Bio import SeqIO
 
-OK_CHR_ORD = {'A': True, 'C': True, 'G': True, 'T': True, 'U': True}
-ALLOWED_NUCL = ['A', 'C', 'G', 'T']
+from source.constants_and_models import ALLOWED_NUCL, OK_CHR_ORD
 
 
 def index_ref(reference_path: str) -> list:
@@ -30,11 +28,11 @@ def index_ref(reference_path: str) -> list:
 
     # check if the reference file already exists
     if absolute_reference_location.with_suffix('.fai').is_file():
-        print('found index ' + str(absolute_reference_location.with_suffix('.fai')))
+        # print('found index ' + str(absolute_reference_location.with_suffix('.fai')))
         index_filename = absolute_reference_location.with_suffix('.fai')
     elif absolute_reference_location.with_suffix(absolute_reference_location.suffix + '.fai').is_file():
-        print('found index ' +
-              str(absolute_reference_location.with_suffix(absolute_reference_location.suffix + '.fai')))
+        # print('found index ' +
+        #       str(absolute_reference_location.with_suffix(absolute_reference_location.suffix + '.fai')))
         index_filename = absolute_reference_location.with_suffix(absolute_reference_location.suffix + '.fai')
     else:
         pass
@@ -100,7 +98,6 @@ def read_ref(ref_path, ref_inds_i, n_handling, n_unknowns=True, quiet=False):
     # TODO convert to SeqIO containers
     # for seq_record in SeqIO.parse(ref_file, "fasta"):
     #     pass
-
 
     ref_file.seek(ref_inds_i[1])
     my_dat = ''.join(ref_file.read(ref_inds_i[2] - ref_inds_i[1]).split('\n'))
@@ -176,6 +173,85 @@ def read_ref(ref_path, ref_inds_i, n_handling, n_unknowns=True, quiet=False):
         print('{0:.3f} (sec)'.format(time.time() - tt))
 
     return my_dat, n_info
+
+
+def find_n_regions(input_sequence: Seq, n_handling: tuple, n_unknowns: bool = False):
+    """
+    Finds N regions in the sequence
+
+    :param input_sequence: Biopython Seq object containing the sequence to scan.
+    :param n_handling: tuple describing the n handling parameters
+    :param n_unknowns: Eliminate unknowns True/False
+    :return:
+    """
+    # data explanation: my_dat[n_atlas[0][0]:n_atlas[0][1]] = solid block of Ns
+    my_dat = input_sequence.tomutable()
+    prev_ni = 0
+    n_count = 0
+    n_atlas = []
+    for i in range(len(my_dat)):
+        # the or part will always be false because n_unknowns is always false
+        if my_dat[i] == 'N' or (n_unknowns and my_dat[i] not in OK_CHR_ORD):
+            if n_count == 0:
+                prev_ni = i
+            n_count += 1
+            if i == len(my_dat) - 1:
+                n_atlas.append((prev_ni, prev_ni + n_count))
+        else:
+            if n_count > 0:
+                n_atlas.append((prev_ni, prev_ni + n_count))
+            n_count = 0
+
+    # handle N base-calls as desired
+    # random is for paired end, ignore for single end allChr is never used.
+    n_info = {'all': [], 'big': [], 'non_N': []}
+    if n_handling[0] == 'random':
+        for region in n_atlas:
+            n_info['all'].extend(region)
+            if region[1] - region[0] <= n_handling[1]:
+                for i in range(region[0], region[1]):
+                    my_dat[i] = random.choice(ALLOWED_NUCL)
+            else:
+                n_info['big'].extend(region)
+
+    # this block of code is never used because 'allChr' is not used.
+    # The current code options are n_handling = ('random', fragment_size) and
+    # n_handling = ('ignore', read_len). Looking at this block it would replace
+    # all N's with whatever base is in the third position of n_handling, so it  would
+    # have to look something like n_handling = ('allChr', fragment_size, 'A').
+    # Bottom line, this part may be deletable
+    elif n_handling[0] == 'allChr' and n_handling[2] in OK_CHR_ORD:
+        for region in n_atlas:
+            n_info['all'].extend(region)
+            if region[1] - region[0] <= n_handling[1]:
+                for i in range(region[0], region[1]):
+                    my_dat[i] = n_handling[2]
+            else:
+                n_info['big'].extend(region)
+
+    elif n_handling[0] == 'ignore':
+        for region in n_atlas:
+            n_info['all'].extend(region)
+            n_info['big'].extend(region)
+    else:
+        print('\nERROR: UNKNOWN N_HANDLING MODE\n')
+        sys.exit(1)
+
+    habitable_regions = []
+    if not n_info['big']:
+        n_info['non_N'] = [(0, len(my_dat))]
+    else:
+        for i in range(0, len(n_info['big']), 2):
+            if i == 0:
+                habitable_regions.append((0, n_info['big'][0]))
+            else:
+                habitable_regions.append((n_info['big'][i - 1], n_info['big'][i]))
+        habitable_regions.append((n_info['big'][-1], len(my_dat)))
+    for n in habitable_regions:
+        if n[0] != n[1]:
+            n_info['non_N'].append(n)
+
+    return n_info
 
 
 def get_all_ref_regions(ref_path, ref_inds, n_handling, save_output=False):
